+++ Filesystem.Akka.Tests/DeleteFileMissing.Tests.cs
@@ -7,9 +7,8 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 namespace Filesystem.Akka.Tests
 {
     [TestClass]
-    public class DeleteFileTests : TestKit
+    public class DeleteFileMissingTests : TestKit
     {
-        private string existingFile;
         private string missingFile;
         private string fileInMissingFolder;
 
@@ -17,29 +16,15 @@ namespace Filesystem.Akka.Tests
         public void Cleanup()
         {
             Shutdown();
-            File.Delete(this.existingFile);
         }
 
         [TestInitialize]
         public void Initialise()
         {
-            this.existingFile = Path.Combine(Path.GetTempPath(), "existing_" + Guid.NewGuid().ToString());
             this.missingFile = Path.Combine(Path.GetTempPath(), "missing_" + Guid.NewGuid().ToString());
             this.fileInMissingFolder = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString(), "missingfolder_" + Guid.NewGuid().ToString());
-
-            File.WriteAllText(this.existingFile, "Test");
         }
 
-        [TestMethod]
-        public void Can_delete_existing_file()
-        {
-            var fs = Sys.ActorOf(Props.Create(() => new Filesystem()));
-            fs.Tell(new DeleteFile(new DeletableFile(this.existingFile)));
-            var result = ExpectMsg<bool>();
-            Assert.IsTrue(result);
-            Assert.IsFalse(File.Exists(this.existingFile));
-        }
-        
         [TestMethod]
         public void Can_delete_missing_file()
         {
